refact: rename to FigureImage
