+++ packages/designmanual/stories/collated-components.jsx
@@ -24,7 +24,7 @@ import Pager from '@ndla/pager';
 import { StoryIntro, StoryBody } from './wrappers';
 import { Center } from './helpers';
 import ArticleLoader from './article/ArticleLoader';
-import FigureWithImage from './article/FigureWithImage';
+import { FigureImage } from './article/FigureImage';
 import { topicList } from '../dummydata/index';
 import MastheadWithTopicMenu, { MastheadWithLogo } from './molecules/mastheads';
 import Tabs, { TabsControlled } from './molecules/tabs';
@@ -410,7 +410,7 @@ storiesOf('Sammensatte moduler', module)
       </LayoutItem>
       <LayoutItem layout="center">
         <h2 className="u-heading">Lisensikoner på enkeltelementer</h2>
-        <FigureWithImage
+        <FigureImage
           alt=""
           src="https://staging.api.ndla.no/image-api/raw/42-45210905.jpg"
           runScripts
